i #199 introduce BTTExtCompletionSuggestion used by TextCompletion
My dearest lover Tom,

it has been a long time,
without you my friend,
and I'll tell you all about it
in this commit message.
I've come a long way,
from where we be began
Oh, I'll tell you all about it
in this commit message.

The tests pass now, turns out text: don't update the suggestions.
Instead we need to call the updateSuggestions function ourself. Also,
it returns a promise. Who could've known. Cost me like 45 minutes of my
precious time. We should look at it tomorrow, maybe we can find a way
to make it prettier.

For now, the text completion widget mostly handles it's data in form of
Suggestion objects, however, the confirmation still is a mess, which I
have not yet found a suitable solution for. There are so many ways to
confirm this thing. The input also CAN handle suggestion-morphs, but it never gets used
that way, mostly because the route on the server is not currently
sending the fontID (i think, didn't investigate...).

My time is running low now. Let me know if you need more information.

May the odds be ever in your favour.
Love,
Justus